+++ src/main/java/dk/kb/image/util/KalturaUtil.java
@@ -47,12 +47,13 @@ public class KalturaUtil {
             baseUrl=baseUrl+"/height/"+height;
         }
 
-        String seek=""; //if seek not defined, this empty string will just be added
-        if ((secondsEndSeek != null && secondsEndSeek>=0)  || (secondsStartSeek != null && secondsStartSeek>=0)) {
+        //Notice Kaltura API has changed. seek=0 will now overrule thumbnail number. Only set if positive.
+        String seek=""; //if seek not defined, this empty string will just be added        
+        if ((secondsEndSeek != null && secondsEndSeek>0)  || (secondsStartSeek != null && secondsStartSeek>0)) {
             seek="?start_sec="+secondsStartSeek+"&end_sec="+secondsEndSeek; 
         }
 
-        thumbnails.setDefault(baseUrl); // Example: https://api.kaltura.nordu.net/p/380/thumbnail/entry_id/0_dtvciomh/width/200/
+        
 
         //This is the sprite version will all thumbnails.
         //Example: https://api.kaltura.nordu.net/p/380/thumbnail/entry_id/0_dtvciomh/width/200/vid_slices/10
@@ -67,6 +68,9 @@ public class KalturaUtil {
         }
         thumbnails.setThumbnails(timeSliceThumbnails);
 
+        //Make #4 defailt (index value 3 is number 4)
+        thumbnails.setDefault(timeSliceThumbnails.get(3)); // Example: https://api.kaltura.nordu.net/p/380/thumbnail/entry_id/0_dtvciomh/width/200/
+        
         return thumbnails;
 
     }