+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Fixed
+- Kaltura thumbnail API was changed. Seek parameter overrules thumbnail number id. Only set seek parameter if they are positive and not 0.
+
+
 ## [5.0.0](https://github.com/kb-dk/ds-image/releases/tag/ds-image-5.0.0) - 2026-06-10
 
 - Removed kaltura dependency.